updated to symfony 2.3. some issues with upgrading, still working through them. updated ugc element so admin can remove others ugc.

+++ src/SkNd/UserBundle/Entity/MemoryWallUGC.php
@@ -78,7 +78,7 @@ class MemoryWallUGC extends MemoryWallContent{
     
     public function removeUpload()
     {
-        if (null !== $this->imagePath) {
+        if (null !== $this->imagePath && file_exists($this->getAbsolutePath())) {
             unlink($this->getAbsolutePath());
             unlink($this->getThumbnailAbsolutePath());
         }
